backend/requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.13.0
PyJWT==1.7.1
PyMySQL==0.10.0
six==1.15.0
Werkzeug==1.0.1
